backend/requirements.txt
djangorestframework==3.14.0
djoser==2.1.0
django-filter==21.1
Pillow==9.5.0
gunicorn==20.0.4
psycopg2-binary==2.9.6
python-dotenv